kill job feature context
