+++ js_modules/dagster-ui/packages/ui-core/src/app/AppProvider.tsx
@@ -35,7 +35,6 @@ import {AssetRunLogObserver} from '../asset-graph/AssetRunLogObserver';
 import {CodeLinkProtocolProvider} from '../code-links/CodeLinkProtocol';
 import {DeploymentStatusProvider, DeploymentStatusType} from '../instance/DeploymentStatusProvider';
 import {InstancePageContext} from '../instance/InstancePageContext';
-import {JobFeatureProvider} from '../pipelines/JobFeatureContext';
 import {WorkspaceProvider} from '../workspace/WorkspaceContext/WorkspaceContext';
 import './blueprint.css';
 
@@ -183,12 +182,10 @@ export const AppProvider = (props: AppProviderProps) => {
                           <CustomConfirmationProvider>
                             <AnalyticsContext.Provider value={analytics}>
                               <InstancePageContext.Provider value={instancePageValue}>
-                                <JobFeatureProvider>
-                                  <LayoutProvider>
-                                    <DagsterPlusLaunchPromotion />
-                                    {props.children}
-                                  </LayoutProvider>
-                                </JobFeatureProvider>
+                                <LayoutProvider>
+                                  <DagsterPlusLaunchPromotion />
+                                  {props.children}
+                                </LayoutProvider>
                               </InstancePageContext.Provider>
                             </AnalyticsContext.Provider>
                           </CustomConfirmationProvider>
