+++ js_modules/dagster-ui/packages/ui-core/src/pipelines/JobFallthroughRoot.oss.tsx
@@ -1,0 +1,6 @@
+import {PipelineOverviewRoot} from './PipelineOverviewRoot';
+import {RepoAddress} from '../workspace/types';
+
+export const JobFallthroughRoot = ({repoAddress}: {repoAddress: RepoAddress}) => {
+  return <PipelineOverviewRoot repoAddress={repoAddress} />;
+};
